Various changes
- Inheritance to simplify architecture of keeping track of modded assets.
- Maps will no longer get added multiple times if mods are enabled and disabled repeatedly.
- All asset fetching has been replaced with a TryGet pattern that should prevent errors when assets are missing.

+++ Assets/KoboldKare/Scripts/Reagents/ScriptableReagent.cs
@@ -37,7 +37,7 @@ public class ScriptableReagent : ScriptableObject {
     public ReagentConsumptionEvent GetConsumptionEvent() => consumptionEvent;
     public Reagent GetReagent( float volume ) {
         return new Reagent() {
-            id = ReagentDatabase.GetID(this),
+            id = (byte)ReagentDatabase.GetID(this),
             volume = volume,
         };
     }
